LPD-25944 Refactor rest dropdowns
