--- modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/components/RestApplicationItem.d.ts
@@ -1,17 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-/// <reference types="react" />
-
-import '../../css/FDSEntries.scss';
-interface IRESTApplicationItemProps {
-	query: string;
-	restApplication: string;
-}
-declare function RESTApplicationItem({
-	query,
-	restApplication,
-}: IRESTApplicationItemProps): JSX.Element;
-export default RESTApplicationItem;
