+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/components/rest/RESTApplicationDropdownMenu.d.ts
@@ -6,9 +6,11 @@
 /// <reference types="react" />
 
 export default function RESTApplicationDropdownMenu({
+	className,
 	onItemClick,
 	restApplications: initialRESTApplications,
 }: {
+	className?: string;
 	onItemClick: Function;
 	restApplications: Array<string>;
 }): JSX.Element;
