+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/components/rest/RESTEndpointDropdownMenu.d.ts
@@ -6,9 +6,11 @@
 /// <reference types="react" />
 
 export default function RESTEndpointDropdownMenu({
+	className,
 	onItemClick,
 	restEndpoints: initialRESTEndpoints,
 }: {
+	className?: string;
 	onItemClick: Function;
 	restEndpoints: Array<string>;
 }): JSX.Element;
