+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/components/rest/RESTSchemaDropdownMenu.d.ts
@@ -6,9 +6,11 @@
 /// <reference types="react" />
 
 export default function RESTSchemaDropdownMenu({
+	className,
 	onItemClick,
 	restSchemas: initialRESTSchemas,
 }: {
+	className?: string;
 	onItemClick: Function;
 	restSchemas: Array<string>;
 }): JSX.Element;
