+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/data_set/filters/modals/selection_filter/source_type/ApiRestApplication.d.ts
@@ -20,9 +20,9 @@ interface IApiRestApplicationModalContentProps {
 	}: {
 		selectedItemKey: string;
 		selectedItemLabel: string;
-		selectedRESTApplication: string | undefined;
-		selectedRESTEndpoint: string | undefined;
-		selectedRESTSchema: string | undefined;
+		selectedRESTApplication: string | null;
+		selectedRESTEndpoint: string | null;
+		selectedRESTSchema: string | null;
 	}) => void;
 	requiredRESTApplicationValidationError: boolean;
 	restApplications: string[];
